+++ apps/tokenbar.js
@@ -56,7 +56,7 @@ export class TokenBar extends Application {
     /** @override */
     getData(options) {
         let css = [
-            ((game.world.system == "dnd5e" || game.world.system == "sw5e" ) && (game.settings.get(game.world.system, 'disableExperienceTracking')) ? 'hidexp' : null),
+            ((game.system.id == "dnd5e" || game.system.id == "sw5e") && (game.settings.get(game.system.id, 'disableExperienceTracking')) ? 'hidexp' : null),
             !game.user.isGM ? "hidectrl" : null
         ].filter(c => !!c).join(" ");
         let pos = this.getPos();
@@ -619,7 +619,7 @@ export class TokenBar extends Application {
         }
     });*/
 
-    //if (game.world.system == "dnd5e") {
+    //if (game.system.id == "dnd5e") {
     //    $('.assign-xp', html).css({ visibility: (game.settings.get('dnd5e', 'disableExperienceTracking') ? 'hidden' : 'visible') });
     //} else {
         //$('.dialog-col', html).hide();